Break require cycle in backgroundNotificationSocket module
Extract the state atom, refresh action, native config helper, and a
read-only enabled atom into a leaf state.ts. The writable enabled atom
moves next to the set action it delegates to, and the token sync action
reads the enabled flag from the leaf module directly, so no file inside
the folder imports index.ts anymore.
