+++ apps/mobile/src/utils/notifications/backgroundNotificationSocket/index.ts
@@ -1,75 +1,30 @@
 import {
   clearCredentials,
-  configure,
   getEnabledPreference,
-  getState,
-  isBatteryExempt,
-  isGmsAvailable,
   requestBatteryExemption,
-  type BackgroundNotificationSocketState,
 } from '@vexl-next/expo-background-notification-socket'
 import {Array, Effect, pipe} from 'effect'
-import {atom, useAtomValue, useSetAtom, type SetStateAction} from 'jotai'
+import {atom, useAtomValue, useSetAtom} from 'jotai'
 import {useCallback, useEffect} from 'react'
 import {Platform} from 'react-native'
-import {apiEnv} from '../../../api'
 import {globalDialogAtom} from '../../../components/GlobalDialog'
 import {vexlNotificationTokenAtom} from '../../../state/notifications/vexlNotificationTokenAtom'
 import {postLoginFlowCompletedScreensAtom} from '../../../state/postLoginOnboarding'
-import getValueFromSetStateActionOfAtom from '../../atomUtils/getValueFromSetStateActionOfAtom'
-import {versionCode} from '../../environment'
 import {translationAtom} from '../../localization/I18nProvider'
 import reportError from '../../reportError'
 import {useAppState} from '../../useAppState'
 import {setBackgroundNotificationSocketEnabledActionAtom} from './setBackgroundNotificationSocketEnabledActionAtom'
+import {
+  backgroundNotificationSocketStateAtom,
+  configureForCurrentSecret,
+  refreshBackgroundNotificationSocketStateActionAtom,
+} from './state'
 
-interface BackgroundNotificationState {
-  // null until the user decided whether to use the background socket
-  readonly enabled: boolean | null
-  readonly batteryExempt: boolean
-  readonly gmsAvailable: boolean
-  readonly socketState: BackgroundNotificationSocketState
-}
-
-// null until read from native
-export const backgroundNotificationSocketStateAtom =
-  atom<BackgroundNotificationState | null>(null)
-
-const readNativeState = async (): Promise<BackgroundNotificationState> => {
-  const [enabled, batteryExempt, gmsAvailable, socketState] = await Promise.all(
-    [getEnabledPreference(), isBatteryExempt(), isGmsAvailable(), getState()]
-  )
-
-  return {enabled, batteryExempt, gmsAvailable, socketState}
-}
-
-export const refreshBackgroundNotificationSocketStateActionAtom = atom(
-  null,
-  async (_get, set): Promise<void> => {
-    if (Platform.OS !== 'android') return
-    try {
-      set(backgroundNotificationSocketStateAtom, await readNativeState())
-    } catch (cause) {
-      reportError(
-        'warn',
-        new Error('Failed to read background notification socket state', {
-          cause,
-        })
-      )
-    }
-  }
-)
-
-export const configureForCurrentSecret = async (
-  secret: string
-): Promise<void> => {
-  await configure({
-    apiUrl: apiEnv.notificationMs,
-    notificationSecret: secret,
-    platform: 'ANDROID',
-    version: versionCode,
-  })
-}
+export {
+  backgroundNotificationSocketEnabledAtom,
+  setBackgroundNotificationSocketEnabledActionAtom,
+} from './setBackgroundNotificationSocketEnabledActionAtom'
+export * from './state'
 
 export const clearBackgroundNotificationSocketActionAtom = atom(
   null,
@@ -109,16 +64,6 @@ export const offerBackgroundNotificationSocketActionAtom = atom(
       })
     )
     await set(setBackgroundNotificationSocketEnabledActionAtom, confirmed)
-  }
-)
-
-export const backgroundNotificationSocketEnabledAtom = atom(
-  (get) => get(backgroundNotificationSocketStateAtom)?.enabled === true,
-  (get, set, update: SetStateAction<boolean>): void => {
-    const enabled = getValueFromSetStateActionOfAtom(update)(
-      () => get(backgroundNotificationSocketStateAtom)?.enabled === true
-    )
-    void set(setBackgroundNotificationSocketEnabledActionAtom, enabled)
   }
 )
